+++ tools/myriota_auth.py
@@ -1,4 +1,18 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2016-2019, Myriota Pty Ltd, All Rights Reserved
+# SPDX-License-Identifier: BSD-3-Clause-Attribution
+#
+# This file is licensed under the BSD with attribution  (the "License"); you
+# may not use these files except in compliance with the License.
+#
+# You may obtain a copy of the License here:
+# LICENSE-BSD-3-Clause-Attribution.txt and at
+# https://spdx.org/licenses/BSD-3-Clause-Attribution.html
+#
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 
 import json
 import getpass
@@ -19,7 +33,7 @@ def post(endpoint, args, region='us-east-1'):
     url = "https://cognito-idp.%s.amazonaws.com/" % region
     response = requests.post(url, json=args, headers=headers)
     result = response.json()
-    if response.status_code != 200:
+    if response.status_code >= 400:
         raise ValueError("; ".join([result["__type"], str(response.status_code), result["message"]]))
     return result